+++ internal/provider/provider.go
@@ -97,9 +97,7 @@ func (p *RootlyProvider) DataSources(ctx context.Context) []func() datasource.Da
 
 // Resources implements provider.Provider.
 func (p *RootlyProvider) Resources(ctx context.Context) []func() resource.Resource {
-	return []func() resource.Resource{
-		NewAlertRouteResource,
-	}
+	return RootlyProviderGeneratedResources
 }
 
 func New(version string) func() provider.Provider {
