+++ internal/provider/provider_gen.go
@@ -1,0 +1,8 @@
+// DO NOT MODIFY: This file is generated by internal/providergen/index.ts. Any changes will be overwritten during the next build.
+package provider
+
+import "github.com/hashicorp/terraform-plugin-framework/resource"
+
+var RootlyProviderGeneratedResources = []func() resource.Resource{
+	NewAlertRouteResource,
+}
